ParkingManagement/requirements.txt
api==0.0.7
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
djangorestframework==3.14.0
idna==3.4
nose==1.3.7
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
